RAG_Advanced/requirements.txt
openai
chromadb
sentence-transformers
python-dotenv
